Init packages
Contains a file with packages/macros I tend to use in everything, along with a "problem" and "subproblem" environment I use on problemsets.  I should really reconfigure this to be a package, but I tried a month ago and had some issues, and didn't want to invest the effort then.